Refactored minerva-specific code into an angular service. Implements delete and edit of annoton. Delete GraphModel.js which mostly got moved into graph.service.js.
