+++ app/inject.tmpl
@@ -13,7 +13,7 @@
     <div
       class="row">
       <div class="col-xs-9 text-center">
-        <h5>{{tvc.title[0].value()}} ({{tvc.model_id}})</h5>
+        <h5>{{tvc.graph.title[0].value()}} ({{tvc.graph.model_id}})</h5>
       </div>
       <div class="col-xs-3">
         <button
